fix apps & update algo_app

+++ packages/apps/catan_dice_app/src/functions/roll-dice.ts
@@ -1,7 +1,6 @@
-import type { uint32 } from '@noshiro/ts-utils';
 import { randInt } from '@noshiro/ts-utils';
 
 export const rollTwoDices = (): [number, number] => [
-  randInt(1 as uint32, 6 as uint32),
-  randInt(1 as uint32, 6 as uint32),
+  randInt(1, 6),
+  randInt(1, 6),
 ];
